% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamcolors.R
\name{get_team_color_vector}
\alias{get_team_color_vector}
\title{Function to extract teamcolors for a specified team}
\usage{
get_team_color_vector(t_name)
}
\arguments{
\item{t_name}{Team Name}
}
\description{
Function to extract teamcolors for a specified team
}
