% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{hubner_cols}
\alias{hubner_cols}
\title{Function to extract hubner colors as hex codes}
\usage{
hubner_cols(...)
}
\arguments{
\item{...}{Character names of hubner_colors}
}
\description{
Function to extract hubner colors as hex codes
}
\author{
Simon Jackson - drsimonj
}
