% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{hubner_pal}
\alias{hubner_pal}
\title{Return function to interpolate a hubner color palette}
\usage{
hubner_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in hubner_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a hubner color palette
}
\author{
Simon Jackson - drsimonj
}
