% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{hubnr_cols}
\alias{hubnr_cols}
\title{Function to extract hubnr colors as hex codes}
\usage{
hubnr_cols(...)
}
\arguments{
\item{...}{Character names of hubnr_colors}
}
\description{
Function to extract hubnr colors as hex codes
}
\author{
Simon Jackson - drsimonj
}
