% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_std_pdf}
\alias{save_std_pdf}
\title{Standard Export Format PDF}
\usage{
save_std_pdf(plot, name, width = 8, height = 6, units = "in")
}
\arguments{
\item{plot}{Plot Object}

\item{name}{File Name - USE PDF FILE ENDING}
}
\description{
Saves PDF with Fonts using Cairo_PDF in a Standard Format
}
