% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_std_png}
\alias{save_std_png}
\title{Standard Export Format PNG - Bitmap}
\usage{
save_std_png(plot, name, width = 8, height = 6, units = "in")
}
\arguments{
\item{plot}{Plot Object}

\item{name}{File Name - USE PNG FILE ENDING}
}
\description{
Saves PNG with Fontsin a Standard Format
}
