% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{scale_color_hubnr}
\alias{scale_color_hubnr}
\title{Color scale constructor for hubnr colors}
\usage{
scale_color_hubnr(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in hubnr_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for hubnr colors
}
\author{
Simon Jackson - drsimonj
}
