% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hubnr}
\alias{theme_hubnr}
\title{hubnR Standard Theme}
\usage{
theme_hubnr(base_size = 12, font = "Roboto")
}
\arguments{
\item{base_size}{Base Size For Plot Text}

\item{font}{Font to Use}
}
\description{
hubnR Standard Theme
}
