% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hubnr_thin}
\alias{theme_hubnr_thin}
\title{hubnR Thin Theme}
\usage{
theme_hubnr_thin(base_size = 12, font = "Roboto Thin")
}
\arguments{
\item{base_size}{Base Size For Plot Text}

\item{font}{Font to Use}
}
\description{
hubnR Thin Theme
}
